split mathics.core.formatter
flake8
