+++ mathics/builtin/box/graphics.py
@@ -41,8 +41,9 @@ from mathics.core.atoms import (
 
 from mathics.core.attributes import A_HOLD_ALL, A_PROTECTED, A_READ_PROTECTED
 from mathics.core.expression import Expression
-from mathics.core.formatter import format_element, lookup_method
+from mathics.core.formatter import lookup_method
 from mathics.core.list import ListExpression
+from mathics.core.makeboxes import format_element
 from mathics.core.symbols import Symbol, SymbolTrue
 from mathics.core.systemsymbols import SymbolAutomatic, SymbolTraditionalForm
