+++ mathics/builtin/files_io/files.py
@@ -29,7 +29,7 @@ from mathics.core.attributes import A_PROTECTED, A_READ_PROTECTED
 from mathics.core.convert.expression import to_expression, to_mathics_list
 from mathics.core.convert.python import from_python
 from mathics.core.expression import BoxError, Expression
-from mathics.core.formatter import format_element, do_format
+from mathics.core.makeboxes import format_element, do_format
 from mathics.core.parser import MathicsFileLineFeeder, parse
 from mathics.core.read import (
     channel_to_stream,
