+++ mathics/builtin/forms.py
@@ -29,8 +29,8 @@ from mathics.core.attributes import A_LOCKED, A_PROTECTED
 
 from mathics.core.element import EvalMixin
 from mathics.core.expression import Expression, BoxError
-from mathics.core.formatter import format_element
 from mathics.core.list import ListExpression
+from mathics.core.makeboxes import format_element
 from mathics.core.symbols import (
     Symbol,
     SymbolFullForm,
