+++ mathics/builtin/makeboxes.py
@@ -27,8 +27,8 @@ from mathics.core.attributes import (
 )
 from mathics.core.element import BaseElement, BoxElementMixin
 from mathics.core.expression import Expression
-from mathics.core.formatter import format_element
 from mathics.core.list import ListExpression
+from mathics.core.makeboxes import format_element
 from mathics.core.number import (
     dps,
     convert_base,
