+++ mathics/builtin/numbers/calculus.py
@@ -58,8 +58,8 @@ from mathics.core.convert.sympy import sympy_symbol_prefix, SympyExpression, fro
 from mathics.core.evaluation import Evaluation
 from mathics.core.evaluators import eval_N
 from mathics.core.expression import Expression
-from mathics.core.formatter import format_element
 from mathics.core.list import ListExpression
+from mathics.core.makeboxes import format_element
 from mathics.core.number import dps, machine_epsilon
 from mathics.core.rules import Pattern
