+++ mathics/builtin/string/operations.py
@@ -40,8 +40,8 @@ from mathics.core.attributes import (
 )
 from mathics.core.convert.python import from_python
 from mathics.core.expression import Expression, string_list
-from mathics.core.formatter import format_element
 from mathics.core.list import ListExpression
+from mathics.core.makeboxes import format_element
 from mathics.core.symbols import (
     Symbol,
     SymbolFalse,
