+++ mathics/core/evaluation.py
@@ -455,7 +455,7 @@ class Evaluation:
         self.stopped = True
 
     def format_output(self, expr, format=None):
-        from mathics.core.formatter import format_element
+        from mathics.core.makeboxes import format_element
 
         if format is None:
             format = self.format
